fixed the while loop bug in getInput
the while loop doesnt exit because the value of the mWaiting was cached
by the system and even when the mWaiting was changed to false by another
method, the cached value isnt updated and the loop doesnt exit.

the keyword "volatile" prevents the variable from being cached. it tells
the system that the variable's value is volatile, like it can change at
any time, hence the system knows not to cache it and so it will be updated
and the problem is solved

Signed-off-by: Nicholas Lum <nicholaslum444@hotmail.com>